ht_submission/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
requests==2.22.0
pytest-django==4.1.0
pytest==6.2.1
pytest-mock==3.5.1
